# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NutriAI — AI Nutrition Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List the current user's food inventory
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add an item to inventory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Inventory item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get an inventory item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Inventory item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Remove an item from inventory
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Grocery Lists ────────────────────────────────────────────────────────────
  /grocery-lists:
    get:
      operationId: listGroceryLists
      tags: [grocery]
      summary: List the current user's grocery lists
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of grocery lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroceryList"
    post:
      operationId: createGroceryList
      tags: [grocery]
      summary: Create a new grocery list
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroceryListInput"
      responses:
        "201":
          description: Grocery list created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryList"
  /grocery-lists/{id}:
    get:
      operationId: getGroceryList
      tags: [grocery]
      summary: Get a grocery list with its items
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Grocery list with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryListWithItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateGroceryList
      tags: [grocery]
      summary: Update a grocery list
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroceryListUpdate"
      responses:
        "200":
          description: Grocery list updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryList"
    delete:
      operationId: deleteGroceryList
      tags: [grocery]
      summary: Delete a grocery list
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /grocery-lists/{listId}/items:
    post:
      operationId: addGroceryListItem
      tags: [grocery]
      summary: Add an item to a grocery list
      security:
        - bearerAuth: []
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroceryListItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryListItem"
  /grocery-lists/{listId}/items/{itemId}:
    patch:
      operationId: updateGroceryListItem
      tags: [grocery]
      summary: Update a grocery list item (e.g. check/uncheck)
      security:
        - bearerAuth: []
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroceryListItemUpdate"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryListItem"
    delete:
      operationId: deleteGroceryListItem
      tags: [grocery]
      summary: Remove an item from a grocery list
      security:
        - bearerAuth: []
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Nutrition ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        foodId:
          type: ["integer", "null"]
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        expiryDate:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - quantity
        - unit
        - createdAt
        - updatedAt
    InventoryItemInput:
      type: object
      properties:
        foodId:
          type: integer
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        expiryDate:
          type: string
          format: date
        notes:
          type: string
      required:
        - name
        - quantity
        - unit
    InventoryItemUpdate:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        expiryDate:
          type: string
          format: date
        notes:
          type: string

    # ── Grocery Lists ─────────────────────────────────────────────────────────
    GroceryListItem:
      type: object
      properties:
        id:
          type: integer
        listId:
          type: integer
        name:
          type: string
        quantity:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        checked:
          type: boolean
        notes:
          type: ["string", "null"]
      required:
        - id
        - listId
        - name
        - checked
    GroceryList:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - createdAt
        - updatedAt
    GroceryListWithItems:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/GroceryListItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - items
        - createdAt
        - updatedAt
    GroceryListInput:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    GroceryListUpdate:
      type: object
      properties:
        name:
          type: string
    GroceryListItemInput:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        notes:
          type: string
      required:
        - name
    GroceryListItemUpdate:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        checked:
          type: boolean
        notes:
          type: string

    # ── Nutrition ─────────────────────────────────────────────────────────────
...