# lib/api-zod/src/generated/types/inventoryItemUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NutriAI — AI Nutrition Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemUpdate {
  name?: string;
  quantity?: number;
  unit?: string;
  expiryDate?: Date;
  notes?: string;
}